Detect basic matches (3 or more horizontally, 3 or more vertically).
Doing this by swapping each candy horizontally with the one next to it and checking if either of the swapped candies make a chain up, down, left or right.
Then swapping each candy vertically with the one below it and checking if either of the swapped candies make a chain up down, left or right.

This also detects L shapes, but need more coding to realize it's an L shape. (it detects 3 good swaps with one of the candies is the same in all 3 then it's an L shape)